EY-2926: BrevDataFacade treng ikkje å ha ansvaret for å fikle med trygdetid, så trekkjer det ut til ei eiga service
