+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/hentinformasjon/BrevdataFacade.kt
@@ -9,8 +9,6 @@ import no.nav.etterlatte.brev.behandling.Beregningsperiode
 import no.nav.etterlatte.brev.behandling.ForenkletVedtak
 import no.nav.etterlatte.brev.behandling.GenerellBrevData
 import no.nav.etterlatte.brev.behandling.PersonerISak
-import no.nav.etterlatte.brev.behandling.Trygdetid
-import no.nav.etterlatte.brev.behandling.Trygdetidsperiode
 import no.nav.etterlatte.brev.behandling.Utbetalingsinfo
 import no.nav.etterlatte.brev.behandling.hentUtbetaltBeloep
 import no.nav.etterlatte.brev.behandling.mapAvdoed
@@ -36,7 +34,7 @@ class BrevdataFacade(
     private val grunnlagKlient: GrunnlagKlient,
     private val beregningKlient: BeregningKlient,
     private val behandlingKlient: BehandlingKlient,
-    private val trygdetidKlient: TrygdetidKlient,
+    private val trygdetidService: TrygdetidService,
 ) {
     suspend fun hentEtterbetaling(
         behandlingId: UUID,
@@ -211,27 +209,5 @@ class BrevdataFacade(
     suspend fun finnTrygdetid(
         behandlingId: UUID,
         brukerTokenInfo: BrukerTokenInfo,
-    ): Trygdetid? {
-        val trygdetidMedGrunnlag = trygdetidKlient.hentTrygdetid(behandlingId, brukerTokenInfo) ?: return null
-
-        val trygdetidsperioder =
-            trygdetidMedGrunnlag.trygdetidGrunnlag.map {
-                Trygdetidsperiode(
-                    datoFOM = it.periodeFra,
-                    datoTOM = it.periodeTil,
-                    land = it.bosted,
-                    opptjeningsperiode = it.beregnet?.aar.toString(),
-                )
-            }
-
-        val beregnetTrygdetid = trygdetidMedGrunnlag.beregnetTrygdetid?.resultat
-        val samlaTrygdetid = beregnetTrygdetid?.samletTrygdetidNorge ?: beregnetTrygdetid?.samletTrygdetidTeoretisk ?: 0
-        val aarTrygdetid = samlaTrygdetid.div(12)
-
-        return Trygdetid(
-            aarTrygdetid = aarTrygdetid,
-            maanederTrygdetid = samlaTrygdetid % 12,
-            perioder = trygdetidsperioder,
-        )
-    }
+    ) = trygdetidService.finnTrygdetid(behandlingId, brukerTokenInfo)
 }
