+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/brev/hentinformasjon/BrevdataFacadeImplTest.kt
@@ -28,19 +28,12 @@ import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
 import no.nav.etterlatte.libs.common.tilbakekreving.Tilbakekreving
 import no.nav.etterlatte.libs.common.toJsonNode
 import no.nav.etterlatte.libs.common.toObjectNode
-import no.nav.etterlatte.libs.common.trygdetid.BeregnetTrygdetidGrunnlagDto
-import no.nav.etterlatte.libs.common.trygdetid.DetaljertBeregnetTrygdetidDto
-import no.nav.etterlatte.libs.common.trygdetid.DetaljertBeregnetTrygdetidResultat
-import no.nav.etterlatte.libs.common.trygdetid.GrunnlagOpplysningerDto
-import no.nav.etterlatte.libs.common.trygdetid.TrygdetidDto
-import no.nav.etterlatte.libs.common.trygdetid.TrygdetidGrunnlagDto
 import no.nav.etterlatte.libs.common.vedtak.Attestasjon
 import no.nav.etterlatte.libs.common.vedtak.VedtakFattet
 import no.nav.etterlatte.libs.common.vedtak.VedtakInnholdDto
 import no.nav.etterlatte.libs.common.vedtak.VedtakNyDto
 import no.nav.etterlatte.libs.common.vedtak.VedtakStatus
 import no.nav.etterlatte.libs.common.vedtak.VedtakType
-import no.nav.etterlatte.libs.testdata.grunnlag.AVDOED_FOEDSELSNUMMER
 import no.nav.etterlatte.libs.testdata.grunnlag.GrunnlagTestData
 import no.nav.etterlatte.libs.testdata.grunnlag.SOEKER_FOEDSELSNUMMER
 import no.nav.etterlatte.token.BrukerTokenInfo
@@ -49,8 +42,6 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.time.LocalDate
-import java.time.Month
 import java.time.YearMonth
 import java.util.UUID
 
@@ -59,7 +50,7 @@ internal class BrevdataFacadeImplTest {
     private val grunnlagKlient = mockk<GrunnlagKlient>()
     private val beregningKlient = mockk<BeregningKlient>()
     private val behandlingKlient = mockk<BehandlingKlient>()
-    private val trygdetidKlient = mockk<TrygdetidKlient>()
+    private val trygdetidService = mockk<TrygdetidService>()
 
     private val service =
         BrevdataFacade(
@@ -67,7 +58,7 @@ internal class BrevdataFacadeImplTest {
             grunnlagKlient,
             beregningKlient,
             behandlingKlient,
-            trygdetidKlient,
+            trygdetidService,
         )
 
     @BeforeEach
@@ -92,7 +83,7 @@ internal class BrevdataFacadeImplTest {
         coEvery { vedtaksvurderingKlient.hentVedtak(any(), any()) } returns opprettBehandlingVedtak()
         coEvery { grunnlagKlient.hentGrunnlag(SAK_ID, BEHANDLING_ID, BRUKERTokenInfo) } returns opprettGrunnlag()
         coEvery { beregningKlient.hentBeregning(any(), any()) } returns opprettBeregning()
-        coEvery { trygdetidKlient.hentTrygdetid(any(), any()) } returns opprettTrygdetid()
+        coEvery { trygdetidService.finnTrygdetid(any(), any()) } returns opprettTrygdetid()
 
         val generellBrevData =
             runBlocking {
@@ -172,7 +163,7 @@ internal class BrevdataFacadeImplTest {
         coEvery { vedtaksvurderingKlient.hentVedtak(any(), any()) } returns opprettBehandlingVedtak()
         coEvery { grunnlagKlient.hentGrunnlag(any(), any(), any()) } returns opprettGrunnlag()
         coEvery { beregningKlient.hentBeregning(any(), any()) } returns opprettBeregning()
-        coEvery { trygdetidKlient.hentTrygdetid(any(), any()) } returns opprettTrygdetid()
+        coEvery { trygdetidService.finnTrygdetid(any(), any()) } returns opprettTrygdetid()
 
         val utbetalingsinfo =
             runBlocking {
@@ -204,7 +195,7 @@ internal class BrevdataFacadeImplTest {
         coEvery { vedtaksvurderingKlient.hentVedtak(any(), any()) } returns opprettBehandlingVedtak()
         coEvery { grunnlagKlient.hentGrunnlag(any(), any(), any()) } returns opprettGrunnlag()
         coEvery { beregningKlient.hentBeregning(any(), any()) } returns opprettBeregningSoeskenjustering()
-        coEvery { trygdetidKlient.hentTrygdetid(any(), any()) } returns opprettTrygdetid()
+        coEvery { trygdetidService.finnTrygdetid(any(), any()) } returns opprettTrygdetid()
 
         val utbetalingsinfo =
             runBlocking {
@@ -216,67 +207,6 @@ internal class BrevdataFacadeImplTest {
 
         coVerify(exactly = 1) { beregningKlient.hentBeregning(any(), any()) }
     }
-
-    @Test
-    fun `henter trygdetid`() {
-        val behandlingId = UUID.randomUUID()
-        coEvery { trygdetidKlient.hentTrygdetid(any(), any()) } returns trygdetidDto(behandlingId)
-        val trygdetid = runBlocking { service.finnTrygdetid(behandlingId, mockk()) }
-        Assertions.assertEquals(3, trygdetid!!.aarTrygdetid)
-        Assertions.assertEquals(6, trygdetid.maanederTrygdetid)
-        Assertions.assertEquals("2", trygdetid.perioder[0].opptjeningsperiode)
-    }
-
-    private fun trygdetidDto(behandlingId: UUID) =
-        TrygdetidDto(
-            id = UUID.randomUUID(),
-            behandlingId = behandlingId,
-            beregnetTrygdetid =
-                DetaljertBeregnetTrygdetidDto(
-                    resultat =
-                        DetaljertBeregnetTrygdetidResultat(
-                            faktiskTrygdetidNorge = null,
-                            faktiskTrygdetidTeoretisk = null,
-                            fremtidigTrygdetidNorge = null,
-                            fremtidigTrygdetidTeoretisk = null,
-                            samletTrygdetidNorge = 42,
-                            samletTrygdetidTeoretisk = null,
-                            prorataBroek = null,
-                            overstyrt = false,
-                        ),
-                    tidspunkt = Tidspunkt.now(),
-                ),
-            trygdetidGrunnlag =
-                listOf(
-                    TrygdetidGrunnlagDto(
-                        id = UUID.randomUUID(),
-                        type = "",
-                        bosted = "Danmark",
-                        periodeFra = LocalDate.of(2020, Month.MARCH, 5),
-                        periodeTil = LocalDate.of(2023, Month.JANUARY, 1),
-                        kilde = null,
-                        beregnet =
-                            BeregnetTrygdetidGrunnlagDto(
-                                dager = 0,
-                                maaneder = 10,
-                                aar = 2,
-                            ),
-                        begrunnelse = null,
-                        poengInnAar = false,
-                        poengUtAar = false,
-                        prorata = false,
-                    ),
-                ),
-            opplysninger =
-                GrunnlagOpplysningerDto(
-                    avdoedDoedsdato = null,
-                    avdoedFoedselsdato = null,
-                    avdoedFylteSeksten = null,
-                    avdoedFyllerSeksti = null,
-                ),
-            overstyrtNorskPoengaar = null,
-            ident = AVDOED_FOEDSELSNUMMER.value,
-        )
 
     private fun opprettBeregning() =
         mockk<BeregningDTO> {
